new function getIframeUrl
